fix the profiles img and some buttons

+++ src/components/profiles/worker/Tools.js
@@ -13,6 +13,7 @@ import {
   Modal,
   Row,
 } from "react-bootstrap";
+import Swal from "sweetalert2";
 function Tools(props) {
   const [userList, setUserList] = useState({});
   const [workerList, setWorkerList] = useState({});
@@ -50,12 +51,48 @@ function Tools(props) {
 
   // :::::: delete from tools ::::::
   const deleteTools = async (indx) => {
-    let res = await axios.delete(`${Api}/worker/tools?index=${indx}`, {
-      headers: {
-        authorization: `Bearer ${token}`,
+
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
       },
-    });
-    setWorkerList(res.data);
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then(async(result) => {
+      if (result.isConfirmed) {
+        let res = await axios.delete(`${Api}/worker/tools?index=${indx}`, {
+          headers: {
+            authorization: `Bearer ${token}`,
+          },
+        });
+        setWorkerList(res.data);
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
+    })
+    
   };
 
   return (
